requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
streamlit>=1.28.0
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.1.0
nltk==3.8.1
spacy>=3.7.2
beautifulsoup4==4.12.2
lxml>=4.9.3
numpy>=1.26.0
scikit-learn>=1.3.2
python-dotenv==1.0.0
aiohttp==3.9.1
httpx==0.25.1